Add extra datastore arguments to vm.create
This in preparation for adding support for creating VMs in datastore
clusters (also known as storage pods). When this lands the VM can be
placed in any of the datastores attached to a datastore cluster. This
means that the datastores for both ISOs and linked disks will need to be
specified explicitly.