Modify govc's vm.create to create VM in one shot
It used to create an empty VM and reconfigure it to add devices.
This change makes it create the VM and its devices in one pass.
This is done in preparation for datastore cluster support, since making
any kind of placement decision only makes sense if VM configuration is
known completely (or storage-related placement decisions would be made
based on a VM without any disks...).